feat: SRS-550: Generic Drawer component
